+++ client/src/container/container.ts
@@ -3,6 +3,7 @@ import { Container } from "inversify";
 import { App } from "../app/App";
 import { LayoutService } from "../services/LayoutService";
 import { ResourceLoaderService } from "../services/ResourceLoaderService";
+import { StatsService } from "../services/StatsService";
 import { WebGLRenderer } from "../services/WebGLRenderer";
 import { LostTreasureScene } from "../scenes/lost-treasure/LostTreasureScene";
 import type { Scene } from "../types/Scene";
@@ -17,6 +18,7 @@ export function createContainer(): Container {
     .bind<ResourceLoaderService>(SERVICE_TYPES.ResourceLoaderService)
     .to(ResourceLoaderService)
     .inSingletonScope();
+  container.bind<StatsService>(SERVICE_TYPES.StatsService).to(StatsService).inSingletonScope();
   container
     .bind<Scene>(SERVICE_TYPES.Scene)
     .to(LostTreasureScene)