+++ client/src/container/serviceTypes.ts
@@ -3,5 +3,6 @@ export const SERVICE_TYPES = {
   LayoutService: Symbol.for("LayoutService"),
   ResourceLoaderService: Symbol.for("ResourceLoaderService"),
   Scene: Symbol.for("Scene"),
+  StatsService: Symbol.for("StatsService"),
   WebGLRenderer: Symbol.for("WebGLRenderer")
 } as const;